# dsh/dsh.openapi.yaml
openapi: 3.1.0
info:
  title: BThwani DSH API
  version: 0.3.0
  description: API contract for client-facing store discovery, catalog management, and checkout flow.
  contact:
    name: BThwani DSH API Governance
    url: https://bthwani.internal/governance/dsh
servers:
  - url: https://api.bthwani.internal/dsh
    description: Production DSH API boundary
  - url: http://localhost:8080
    description: Local DSH backend runtime
...
paths:
...
  /stores/{store_id}/products:
    post:
      tags:
        - Product Identity
      summary: Create product for a store
      description: >
        Creates a new product identity record for the given store.
        The product starts in `partner_submitted` approval status and enters the catalog review pipeline.
        WLT boundary: base_price_label is a display label only — no financial mutation occurs inside DSH.
      operationId: createProduct
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRecord'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Product Identity
      summary: List products for a store
      description: Returns all product records for the given store, ordered by creation date descending.
      operationId: listProducts
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
        - name: approval_status
          in: query
          required: false
          schema:
            type: string
          description: Filter products by approval status (e.g. catalog_adopted)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProductsResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags:
        - Product Identity
      summary: Get a single product
      description: Returns one product identity record by ID, including DSH catalog status and display-only price label.
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRecord'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Product Identity
      summary: Update product identity
      description: >
        Updates product name, SKU, GTIN, barcode, description, base_price_label, or category_id.
        All fields are optional — only provided fields are updated.
        WLT boundary: base_price_label is a display label only.
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Updated product record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRecord'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
               schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── J-002 / DSH-SLICE-002B: Category Structure ─────────────────────────────
...
  /media:
    post:
      tags:
        - Product Media
      summary: Upload product media record
      description: Creates/attaches a new product media record validated against the local media manifest.
      operationId: uploadProductMedia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadProductMediaRequest'
      responses:
        '201':
          description: Media created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductMediaRecord'
        '400':
          description: Validation error or invalid media key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /media/{id}:
    delete:
      tags:
        - Product Media
      summary: Delete product media
      description: Deletes/removes a product media record.
      operationId: deleteProductMedia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Media deleted
        '404':
          description: Media record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── J-002 / DSH-SLICE-002D: Partner Local Overrides ────────────────────────
  /stores/{store_id}/catalog-overrides:
    patch:
      tags:
        - Partner Local Overrides
      summary: Apply catalog overrides for a store
      description: >
        Applies local overrides (price, stock, availability) for products in the given store's catalog.
        These local overrides do not mutate the central product definitions.
      operationId: updateCatalogOverrides
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCatalogOverridesRequest'
      responses:
        '200':
          description: Overrides updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCatalogOverridesResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── J-002 / DSH-SLICE-002E: Approval Workflow ──────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional bearer token for personalized store discovery (e.g. retrieving user's favorite stores).
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    # ─── J-002 / DSH-SLICE-002A: Product Identity ─────────────────────────────
    ProductRecord:
      type: object
      required:
        - id
        - store_id
        - name
        - base_price_label
        - approval_status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        store_id:
          type: string
        name:
          type: string
        sku:
          type: string
        gtin:
          type: string
        barcode:
          type: string
        description:
          type: string
        base_price_label:
          type: string
          description: Display label only — no financial mutation. WLT owns price semantics.
        category_id:
          type: string
        approval_status:
          type: string
          enum:
            - field_draft
            - partner_submitted
            - partner_review
            - partner_approved
            - marketing_review
            - marketing_approved
            - catalog_adopted
            - client_visible
            - needs_fix
            - rejected
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: '#/components/schemas/ProductMediaRecord'
        price_override:
          type: string
          description: Display-only local price override.
        stock_override:
          type: integer
          description: Local stock count override.
        available_override:
          type: boolean
          description: Local availability override.
    CreateProductRequest:
      type: object
      required:
        - name
        - base_price_label
      properties:
        name:
          type: string
        sku:
          type: string
        gtin:
          type: string
        barcode:
          type: string
        description:
          type: string
        base_price_label:
          type: string
        category_id:
          type: string
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        gtin:
          type: string
        barcode:
          type: string
        description:
          type: string
        base_price_label:
          type: string
        category_id:
          type: string
    ListProductsResponse:
      type: object
      required:
        - products
        - pagination
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ─── J-002 / DSH-SLICE-002B: Category Structure ─────────────────────────────
...
    ProductMediaRecord:
      type: object
      required:
        - id
        - product_id
        - media_key
        - url
        - created_at
      properties:
        id:
          type: string
        product_id:
          type: string
        media_key:
          type: string
        url:
          type: string
        created_at:
          type: string
          format: date-time
    UploadProductMediaRequest:
      type: object
      required:
        - product_id
        - media_key
      properties:
        product_id:
          type: string
        media_key:
          type: string

    # ─── J-002 / DSH-SLICE-002D: Partner Local Overrides ────────────────────────
    UpdateCatalogOverridesRequest:
      type: object
      required:
        - overrides
      properties:
        overrides:
          type: array
          items:
            $ref: '#/components/schemas/CatalogOverrideInput'
    CatalogOverrideInput:
      type: object
      required:
        - product_id
      properties:
        product_id:
          type: string
        price_override:
          type: string
          description: Partner-specific price override label (display-only, WLT boundary).
        stock_override:
          type: integer
          description: Partner-specific inventory stock override count.
        available_override:
          type: boolean
          description: Toggle availability override for the product.
    UpdateCatalogOverridesResponse:
      type: object
      required:
        - store_id
        - updated_count
        - overrides
      properties:
        store_id:
          type: string
        updated_count:
          type: integer
        overrides:
          type: array
          items:
            $ref: '#/components/schemas/CatalogOverrideRecord'
    CatalogOverrideRecord:
      type: object
      required:
        - store_id
        - product_id
        - updated_at
      properties:
        store_id:
          type: string
        product_id:
          type: string
        price_override:
          type: string
        stock_override:
          type: integer
        available_override:
          type: boolean
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

    # ─── J-003 / Checkout Schemas ─────────────────────────────────────────────
...